pyproject.toml
[project]
name = "dpy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dagger-io>=0.18.4",
]
